Redefine split_dataset
Instead of using the random_split function to filter from the dataset, the filtering is now at csv level.

To get a test dataset, we need to take it from the train dataset. If we load the dataset and then split it with random_split, the test dataset will also contain transformations. To avoid that, now there are three csv for the three datasets.